Move unicode-related API to the correct header file.
